seperated out functions and api files from the tests file
