+++ product.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <link rel="stylesheet" href="src/style.css">
-    <link rel="stylesheet" href="src/product-detail.css">
+    <link rel="stylesheet" href="src/product.css">
     <title>Document</title>
 </head>
 <body>
